removing unit test dependency on other unit test; comments cleanup
