+++ HIRS_AttestationCA/src/test/java/hirs/attestationca/persist/entity/userdefined/report/DeviceInfoReportTest.java
@@ -7,17 +7,12 @@ import hirs.attestationca.persist.entity.userdefined.info.NetworkInfo;
 import hirs.attestationca.persist.entity.userdefined.info.HardwareInfo;
 import hirs.attestationca.persist.entity.userdefined.info.FirmwareInfo;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-
 /**
- * DeviceInfoReportTest is a unit test class for DeviceInfoReports.
+ * Unit test class for DeviceInfoReports.
  */
 public class DeviceInfoReportTest extends AbstractUserdefinedEntityTest {
     private final NetworkInfo networkInfo = createTestNetworkInfo();
@@ -25,8 +20,6 @@ public class DeviceInfoReportTest extends AbstractUserdefinedEntityTest {
     private final FirmwareInfo firmwareInfo = createTestFirmwareInfo();
     private final HardwareInfo hardwareInfo = createTestHardwareInfo();
     private final TPMInfo tpmInfo = createTPMInfo();
-
-    private static final Logger LOGGER = LogManager.getLogger(DeviceInfoReportTest.class);
 
     private static final String EXPECTED_CLIENT_VERSION = "Test.Version";
 
@@ -95,67 +88,5 @@ public class DeviceInfoReportTest extends AbstractUserdefinedEntityTest {
         assertEquals(hardwareInfo, deviceInfoReport.getHardwareInfo());
         assertEquals(tpmInfo, deviceInfoReport.getTpmInfo());
         assertEquals(EXPECTED_CLIENT_VERSION, deviceInfoReport.getClientApplicationVersion());
-    }
-
-    /**
-     * Creates a DeviceInfoReport instance usable for testing.
-     *
-     * @return a test DeviceInfoReport
-     */
-    public static DeviceInfoReport getTestReport() {
-        return new DeviceInfoReport(
-                createTestNetworkInfo(), createTestOSInfo(), createTestFirmwareInfo(),
-                createTestHardwareInfo(), createTPMInfo()
-        );
-    }
-
-    /**
-     * Creates a test instance of NetworkInfo.
-     *
-     * @return network information for a fake device
-     */
-    public static NetworkInfo createTestNetworkInfo() {
-        try {
-            final String hostname = "test.hostname";
-            final InetAddress ipAddress =
-                    InetAddress.getByAddress(new byte[] {127, 0, 0, 1});
-            final byte[] macAddress = new byte[] {11, 22, 33, 44, 55, 66};
-            return new NetworkInfo(hostname, ipAddress, macAddress);
-
-        } catch (UnknownHostException e) {
-            LOGGER.error("error occurred while creating InetAddress");
-            return null;
-        }
-    }
-
-    /**
-     * Creates a test instance of OSInfo.
-     *
-     * @return OS information for a fake device
-     */
-    public static OSInfo createTestOSInfo() {
-        return new OSInfo("test os name", "test os version", "test os arch",
-                "test distribution", "test distribution release");
-    }
-
-    /**
-     * Creates a test instance of FirmwareInfo.
-     *
-     * @return Firmware information for a fake device
-     */
-    public static FirmwareInfo createTestFirmwareInfo() {
-        return new FirmwareInfo("test bios vendor", "test bios version", "test bios release date");
-    }
-
-    /**
-     * Creates a test instance of HardwareInfo.
-     *
-     * @return Hardware information for a fake device
-     */
-    public static HardwareInfo createTestHardwareInfo() {
-        return new HardwareInfo("test manufacturer", "test product name", "test version",
-                "test really long serial number with many characters", "test really long chassis "
-                + "serial number with many characters",
-                "test really long baseboard serial number with many characters");
     }
 }
